refactor: single getRowMatch shared by settings sections
general-settings.tsx and performance-settings.tsx each carried a
byte-identical getRowMatch helper. Extracts one copy into the settings-pane
search module (colocated with the RowMatch/SectionMatchView types it
operates on) and imports it from both components.
